+++ packages/zosmf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OSMF SDK package will be documented in this file.
 
+## Unreleased (branch: feat/secure-cert-support)
+
+- Enhancement: Added ZOSMF session option constants and profile fields to support `certAccount` and `certKeyAccount`. Files changed: `packages/zosmf/src/ZosmfSession.ts`, `packages/zosmf/src/constants/Zosmf.profile.ts`.
+
 ## `8.27.4`
 
 - BugFix: Updated minimum supported version of Node from 18 to 20. Added Node 24 support. [#2616](https://github.com/zowe/zowe-cli/pull/2616)